pyproject.toml
[tool.poetry]
name = "fairy-note"
version = "0.1.0"
description = ""
authors = ["cyanife <cyanife@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
aerich = "^0.4.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.16.19"}
uvloop = "^0.14.0"
fastapi-users = {extras = ["tortoise-orm"], version = "^4.0.0"}
python-rapidjson = "^1.0"
ciso8601 = "^2.1.3"
uvicorn = "^0.13.3"
ujson = "^4.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"
flake8 = "^3.8.4"
black = "^20.8b1"
aiomysql = "^0.0.21"
rope = "^0.18.0"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
